requirements.txt
langchain_ollama
langchain_core.prompts
langchain_text_splitters
kagglehub
chromadb
pandas